+++ GUGUENGINE/Level/Level11.h
@@ -93,8 +93,8 @@ private:
 	GLint textureQuitUI3p;
 	GLint textureOptionUI3p;
 	GLint textureRestartUI3p;
-	float first;
-	float last;
+	double first;
+	double last;
 	bool failS[2] = { false };
 	bool UI[5] = { false };
